\name{tabbedPlots-package}
\alias{tabbedPlots-package}
\docType{package}
\title{
  Creates a window to hold multiple plots in tabs
}
\description{
  Creates a window to hold multiple plots in tabs, using the RGtk2 GUI system.
}
\author{
  Lars Hansen and Tony Plate.

Maintainer: Tony Plate <tplatet@acm.org>
}
\keyword{ package }
