\name{tabbedPlots.options}
\alias{tabbedPlots.options}
\title{ Set options for tabbedPlots }
\description{
  Set options for tabbedPlots
}
\usage{
tabbedPlots.options(...)
}
\arguments{
  \item{\dots}{ Options can be any of "active", "debug", "height",
    "reusePar", "width" and should be specified as named arguments}
}
\details{
  Set options for tabbedPlots
}
\value{
  A list of the options before changes.  If no arguments were supplied, all
  user-changable options are returns, if some arguments were supplied,
  only prior values for those options are returned.
}
\keyword{ graphics }
