\name{tabbedPlots.prePlotNewHook}
\alias{tabbedPlots.postPlotNewHook}
\alias{tabbedPlots.prePlotNewHook}
\alias{tabbedPlots.parHook}
\alias{tabbedPlots.layoutHook}
\title{ Internal hook functions }
\description{
  Functions to be called the hooks before and after creating a new plot.
}
\usage{
tabbedPlots.prePlotNewHook()
tabbedPlots.postPlotNewHook()
tabbedPlots.parHook()
tabbedPlots.layoutHook()
}
\details{
  These functions are intended to be called before and after creating a
  new plot.  These functions should not need to be called by the user in
  normal usage.
}
\keyword{ graphics }
