\name{tabbedPlots.warning}
\alias{tabbedPlots.warning}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Fetch the most recent warning }
\description{
  Fetch the most recent warning stored in tabbedPlots and reset it to be empty.
}
\usage{
tabbedPlots.warning()
}
\value{
  Character string.
}
\note{ ~~further notes~~
  The warning is erased.
}
\keyword{ graphics }
